+++ src/services/users/types.ts
@@ -702,10 +702,19 @@ export interface UsersService {
    */
   findAll(options?: OptionsWithRql): Promise<User[]>;
   /**
-   * Find First
-   * @param rql an optional rql string
-   * @returns the first element found
+   * Request a list of all users
+   *
+   * Do not pass in an rql with limit operator!
+   *
+   * Permission | Scope | Effect
+   * - | - | -
+   * none | `patient enlistment` | See a limited set of fields of the staff members (of the groups where you are enlisted as a patient)
+   * none | `staff enlistment` | See a limited set of fields of all patients and staff members (of the groups where you are enlisted as staff member)
+   * `VIEW_USER` | `global` | See all fields of all users
+   * @param rql Add filters to the requested list.
+   * @returns User[]
    */
+  findAllGenerator(options?: OptionsWithRql): AsyncGenerator<User[], [], void>;
   findFirst(options?: { rql?: RQLString }): Promise<User>;
   /**
    * @deprecated